wip refactoring OCR Service and Tesseract setup
