--- Receiptly.Test/UnitTest1.cs
@@ -1,9 +1,0 @@
-namespace Receiptly.Test;
-
-public class UnitTest1
-{
-    [Fact]
-    public void Test1()
-    {
-    }
-}
